yumyumapp/requirements.txt
Django==4.2.25
djangorestframework==3.16.1
django-cors-headers==4.9.0
